Renamed fw to sniffer.
